+++ jans-cli-tui/cli_tui/plugins/010_auth_server/edit_client_dialog.py
@@ -1,3 +1,5 @@
+import copy
+
 from collections import OrderedDict
 from functools import partial
 from typing import Any
@@ -32,6 +34,9 @@ from wui_components.jans_date_picker import DateSelectWidget
 from utils.utils import DialogUtils
 from wui_components.jans_vetrical_nav import JansVerticalNav
 from wui_components.jans_label_container import JansLabelContainer
+from wui_components.jans_label_widget import JansLabelWidget
+
+
 
 from view_uma_dialog import ViewUMADialog
 import threading
@@ -43,11 +48,11 @@ from prompt_toolkit.eventloop import get_event_loop
 import asyncio
 
 import json
-
 ERROR_GETTING_CLIENTS = _("Error getting clients")
 ATTRIBUTE_SCHEMA_PATH = '#/components/schemas/ClientAttributes'
 URL_SUFFIX_FORMATTER = 'inum:{}'
-
+INTROSPECTION_ALG_PROPERTIES = ('introspectionSignedResponseAlg', 'introspectionEncryptedResponseAlg', 'introspectionEncryptedResponseEnc')
+APP = get_app()
 
 class EditClientDialog(JansGDialog, DialogUtils):
     """The Main Client Dialog that contain every thing related to The Client
@@ -99,7 +104,7 @@ class EditClientDialog(JansGDialog, DialogUtils):
         """method to invoked when saving the dialog (Save button is pressed)
         """
 
-        current_data = self.data
+        current_data = copy.deepcopy(self.data)
         self.data = self.make_data_from_dialog()
         self.data['disabled'] = not self.data['disabled']
 
@@ -111,7 +116,7 @@ class EditClientDialog(JansGDialog, DialogUtils):
             'requestUris',
             'claimRedirectUris',
         ):
-            if self.data[list_key]:
+            if list_key in self.data:
                 self.data[list_key] = self.data[list_key].splitlines()
 
         self.data['scopes'] = [item[0] for item in self.client_scopes.entries]
@@ -123,39 +128,41 @@ class EditClientDialog(JansGDialog, DialogUtils):
             self.data['rptAsJwt'] = self.data['rptAsJwt'] == 'jwt'
 
         self.data['attributes'] = {}
-        self.data['attributes'] = {
-            'redirectUrisRegex': self.data['redirectUrisRegex']}
-        self.data['attributes'] = {'parLifetime': self.data['parLifetime']}
-        self.data['attributes'] = {'requirePar': self.data['requirePar']}
+        self.data['attributes']['redirectUrisRegex'] = self.data.pop('redirectUrisRegex')
+        self.data['attributes']['parLifetime'] = self.data.pop('parLifetime')
+        self.data['attributes']['requirePar'] = self.data.pop('requirePar')
+        #self.data['attributes']['requirePkce'] = self.data.pop('requirePkce', False)
 
         for list_key in (
             'backchannelLogoutUri',
             'additionalAudience',
             'rptClaimsScripts',
             'spontaneousScopeScriptDns',
-            'tlsClientAuthSubjectDn',
-            'spontaneousScopes',
-            'updateTokenScriptDns',
-            'postAuthnScripts',
-            'introspectionScripts',
-            'ropcScripts',
-            'consentGatheringScripts',
-
         ):
-            if self.data[list_key]:
-                self.data['attributes'][list_key] = self.data[list_key].splitlines()
+            if list_key in self.data:
+                self.data['attributes'][list_key] = self.data.pop(list_key,'').splitlines()
 
-        for list_key in (
+        for key in (
             'runIntrospectionScriptBeforeJwtCreation',
             'backchannelLogoutSessionRequired',
             'jansDefaultPromptLogin',
             'allowSpontaneousScopes',
+            'tlsClientAuthSubjectDn',
         ):
-            if self.data[list_key]:
-                self.data['attributes'][list_key] = self.data[list_key]
+            if key in self.data:
+                self.data['attributes'][key] = self.data.pop(key)
+
+        for scr_var in self.scripts_widget_dict:
+            values = self.scripts_widget_dict[scr_var].get_values()
+            if values:
+                self.data['attributes'][scr_var] = values
 
         self.data['displayName'] = self.data['clientName']
         self.data['attributes']['jansAuthorizedAcr'] = self.data.pop('jansAuthorizedAcr')
+
+        for intro_attr in INTROSPECTION_ALG_PROPERTIES:
+            if intro_attr in self.data:
+                self.data['attributes'][intro_attr] = self.data.pop(intro_attr)
 
         cfr = self.check_required_fields()
 
@@ -169,6 +176,9 @@ class EditClientDialog(JansGDialog, DialogUtils):
         for prop in current_data:
             if prop not in self.data:
                 self.data[prop] = current_data[prop]
+
+        # remove authenticationMethod, it is read only
+        self.data.pop('authenticationMethod', None)
 
         if self.save_handler:
             self.save_handler(self)
@@ -247,6 +257,9 @@ class EditClientDialog(JansGDialog, DialogUtils):
 
         client_secret_next_widget.handler = partial(change_view_hide, client_secret_widget)
 
+        #require_pkce = self.data.get('attributes', {}).get('redirectUrisRegex')
+        #if require_pkce is None:
+        #    require_pkce = self.myparent.app_configuration.get('requirePkce', False)
 
         basic_tab_widgets = [
             self.myparent.getTitledText(
@@ -335,6 +348,13 @@ class EditClientDialog(JansGDialog, DialogUtils):
                 checked=self.data.get('trustedClient'),
                 jans_help=self.myparent.get_help_from_schema(schema, 'trustedClient'),
                 style=cli_style.check_box),
+
+            #self.myparent.getTitledCheckBox(
+            #    _("Require PKCE"),
+            #    name='requirePkce',
+            #    checked=require_pkce,
+            #    jans_help=self.myparent.get_help_from_schema(schema, 'requirePkce'),
+            #    style=cli_style.check_box),
 
             self.myparent.getTitledRadioButton(
                 _("Application Type"),
@@ -519,7 +539,32 @@ class EditClientDialog(JansGDialog, DialogUtils):
         ], width=D(), style=cli_style.tabs
         )
 
-        self.tabs['SoftwareInfo'] = HSplit([
+        self.tabs['Software Info'] = HSplit([
+
+
+            self.myparent.getTitledText(
+                _("Client URI"),
+                name='clientUri',
+                value=self.data.get('clientUri', ''),
+                jans_help=self.myparent.get_help_from_schema(
+                    schema, 'redirectUris'),
+                style=cli_style.titled_text),
+
+            self.myparent.getTitledText(
+                _("Policiy URI"),
+                name='policyUri',
+                value=self.data.get('policyUri', ''),
+                jans_help=self.myparent.get_help_from_schema(
+                    schema, 'policyUri'),
+                style=cli_style.titled_text),
+
+            self.myparent.getTitledText(
+                _("Logo URI"),
+                name='logoUri',
+                value=self.data.get('logoUri', ''),
+                jans_help=self.myparent.get_help_from_schema(
+                    schema, 'logoUri'),
+                style=cli_style.titled_text),
 
             self.myparent.getTitledText(_("Contacts"),  # height =3 insted of the <+> button
                                         name='contacts',
@@ -723,6 +768,14 @@ class EditClientDialog(JansGDialog, DialogUtils):
                  'request_object_encryption_alg_values_supported'),
                 (_("Request Object Enc for Encryption"), 'requestObjectEncryptionEnc',
                  'request_object_encryption_enc_values_supported'),
+
+                 (_("Introspection Signed Response Alg "), 'introspectionSignedResponseAlg',
+                 'id_token_signing_alg_values_supported'),
+                (_("Introspection Encrypted Response Alg"), 'introspectionEncryptedResponseAlg',
+                 'id_token_encryption_alg_values_supported'),
+                (_("Introspection Encrypted Response Enc"), 'introspectionEncryptedResponseEnc',
+                 'id_token_encryption_enc_values_supported'),
+
         ):
 
             self.drop_down_select_first.append(swagger_key)
@@ -730,12 +783,14 @@ class EditClientDialog(JansGDialog, DialogUtils):
             values = [(alg, alg) for alg in self.myparent.cli_object.openid_configuration.get(
                 openid_key, [])]
 
+            value = self.data.get('attributes', {}).get(swagger_key) if swagger_key in INTROSPECTION_ALG_PROPERTIES else self.data.get(swagger_key)
+
             encryption_signing.append(self.myparent.getTitledWidget(
                 title,
                 name=swagger_key,
                 widget=DropDownWidget(
                     values=values,
-                    value=self.data.get(swagger_key)
+                    value=value
                 ),
                 jans_help=self.myparent.get_help_from_schema(
                     schema, swagger_key),
@@ -854,7 +909,7 @@ class EditClientDialog(JansGDialog, DialogUtils):
                 _("TLS Subject DN"),
                 name='tlsClientAuthSubjectDn',
                 value='\n'.join(self.data.get('attributes', {}).get(
-                                'tlsClientAuthSubjectDn', [])),
+                                'tlsClientAuthSubjectDn') or []),
                 height=3, style=cli_style.check_box,
                 jans_help=self.myparent.get_help_from_schema(
                     self.myparent.cli_object.get_schema_from_reference(
@@ -876,79 +931,37 @@ class EditClientDialog(JansGDialog, DialogUtils):
         ], width=D(), style=cli_style.tabs
         )
 
-        self.tabs['Client Scripts'] = HSplit([
+
+        self.scripts_widget_dict = OrderedDict()
 
 
-            self.myparent.getTitledText(_("Spontaneous Scopes"),
-                                        name='spontaneousScopes',
-                                        value='\n'.join(self.data.get(
-                                            'attributes', {}).get('spontaneousScopes', [])),
-                                        height=3,
-                                        jans_help=self.myparent.get_help_from_schema(
-                self.myparent.cli_object.get_schema_from_reference(
-                    '', ATTRIBUTE_SCHEMA_PATH),
-                'spontaneousScopes'),
-                style=cli_style.check_box),
+        list_of_scripts = (
+                ("Spontaneous Scopes", 'spontaneousScopes', ['spontaneous_scope', 'uma_claims_gathering', 'uma_rpt_policy']),
+                ("Update Token", 'updateTokenScriptDns', ['update_token']),
+                ("Post Authn", 'postAuthnScripts', ['post_authn']),
+                ("Introspection", 'introspectionScripts', ['introspection', 'persistence_extension', 'person_authentication']),
+                ("Password Grant", 'ropcScripts', ['resource_owner_password_credentials', 'scim']),
+                ("OAuth Consent", 'consentGatheringScripts', ['application_session', 'authorization_challenge',  'cache_refresh', 'ciba_end_user_notification', 'client_registration', 'config_api_auth',  'consent_gathering', 'discovery', 'dynamic_scope', 'end_session', 'id_generator', 'idp'])
+                )
 
-            self.myparent.getTitledText(_("Update Token"),
-                                        name='updateTokenScriptDns',
-                                        value='\n'.join(self.data.get('attributes', {}).get(
-                                            'updateTokenScriptDns', [])),
-                                        height=3,
-                                        jans_help=self.myparent.get_help_from_schema(
-                self.myparent.cli_object.get_schema_from_reference(
-                    '', ATTRIBUTE_SCHEMA_PATH),
-                'updateTokenScriptDns'),
-                style=cli_style.check_box),
 
-            self.myparent.getTitledText(_("Post Authn"),
-                                        name='postAuthnScripts',
-                                        value='\n'.join(self.data.get(
-                                            'attributes', {}).get('postAuthnScripts', [])),
-                                        height=3,
-                                        jans_help=self.myparent.get_help_from_schema(
-                self.myparent.cli_object.get_schema_from_reference(
-                    '', ATTRIBUTE_SCHEMA_PATH),
-                'postAuthnScripts'),
-                style=cli_style.check_box),
+        for title, script_var, script_types in list_of_scripts:
+            scripts_data = []
+            for scr in common_data.enabled_scripts:
+                if scr['scriptType'] in script_types:
+                    scripts_data.append((scr['dn'], scr['name']))
 
-            self.myparent.getTitledText(_("Introspection"),
-                                        name='introspectionScripts',
-                                        value='\n'.join(self.data.get('attributes', {}).get(
-                                            'introspectionScripts', [])),
-                                        height=3,
-                                        jans_help=self.myparent.get_help_from_schema(
-                self.myparent.cli_object.get_schema_from_reference(
-                    '', ATTRIBUTE_SCHEMA_PATH),
-                'introspectionScripts'),
-                style=cli_style.check_box),
+            self.scripts_widget_dict[script_var] = JansLabelWidget(
+                        title = _(title), 
+                        values = self.data.get('attributes', {}).get(script_var, []),
+                        data = scripts_data
+                        )
 
-            self.myparent.getTitledText(_("Password Grant"),
-                                        name='ropcScripts',
-                                        value='\n'.join(self.data.get(
-                                            'attributes', {}).get('ropcScripts', [])),
-                                        height=3,
-                                        style=cli_style.check_box,
-                                        jans_help=self.myparent.get_help_from_schema(
-                self.myparent.cli_object.get_schema_from_reference(
-                    '', ATTRIBUTE_SCHEMA_PATH),
-                'ropcScripts'),
-            ),
 
-            self.myparent.getTitledText(_("OAuth Consent"),
-                                        name='consentGatheringScripts',
-                                        value='\n'.join(self.data.get('attributes', {}).get(
-                                            'consentGatheringScripts', [])),
-                                        height=3,
-                                        jans_help=self.myparent.get_help_from_schema(
-                self.myparent.cli_object.get_schema_from_reference(
-                    '', ATTRIBUTE_SCHEMA_PATH),
-                'consentGatheringScripts'),
-                style=cli_style.check_box),
-        ], width=D(), style=cli_style.tabs
-        )
+        self.tabs['Client Scripts'] = HSplit(list(self.scripts_widget_dict.values()), width=D(), style=cli_style.tabs)
 
         self.left_nav = list(self.tabs.keys())[0]
+
 
     def scope_exists(self, scope_dn: str) -> bool:
         for item_id, item_label in self.client_scopes.entries:
@@ -1202,7 +1215,7 @@ class EditClientDialog(JansGDialog, DialogUtils):
                 ),
             ])
 
-            get_app().invalidate()
+            APP.invalidate()
             self.myparent.layout.focus(self.uma_resources)
 
         else: