+++ jans-config-api/server/src/main/java/io/jans/configapi/rest/ApiApplication.java
@@ -6,7 +6,7 @@
 
 package io.jans.configapi.rest;
 
-import io.jans.configapi.configuration.ObjectMapperContextResolver;
+import io.jans.configapi.core.rest.BaseApiApplication;
 import io.jans.configapi.rest.resource.auth.*;
 import io.jans.configapi.util.ApiAccessConstants;
 import io.jans.configapi.rest.health.ApiHealthCheck;
@@ -21,16 +21,19 @@ import io.swagger.v3.oas.annotations.servers.*;
 import java.util.HashSet;
 import java.util.Set;
 import jakarta.ws.rs.ApplicationPath;
-import jakarta.ws.rs.core.Application;
 
 /**
  * @author Mougang T.Gasmyr
  *
  */
 @ApplicationPath("/api/v1")
-@OpenAPIDefinition(info = @Info(title = "Jans Config API", version = "1.0.0", contact = @Contact(name = "Gluu Support", url = "https://support.gluu.org", email = "xxx@gluu.org"),
+@OpenAPIDefinition(info = @Info(title = "Jans Config API", contact =
+@Contact(name = "Contact", url = "https://github.com/JanssenProject/jans/discussions"),
 
-        license = @License(name = "Apache 2.0", url = "https://github.com/JanssenProject/jans/blob/main/LICENSE")),
+        license = @License(name = "License", url = "https://github" +
+                ".com/JanssenProject/jans/blob/main/LICENSE"),
+
+        version = "OAS Version"),
 
         tags = { @Tag(name = "Attribute"), @Tag(name = "Default Authentication Method"),
                 @Tag(name = "Cache Configuration"), @Tag(name = "Cache Configuration – Memcached"),
@@ -46,7 +49,7 @@ import jakarta.ws.rs.core.Application;
                 @Tag(name = "Auth Server Health - Check"), @Tag(name = "Plugins"),
                 @Tag(name = "Configuration – Config API"), @Tag(name = "Client Authorization")},
 
-        servers = { @Server(url = "https://jans.io/", description = "The Jans server") })
+        servers = { @Server(url = "https://jans.local.io", description = "The Jans server") })
 
 @SecurityScheme(name = "oauth2", type = SecuritySchemeType.OAUTH2, flows = @OAuthFlows(clientCredentials = @OAuthFlow(tokenUrl = "https://{op-hostname}/.../token", scopes = {
         @OAuthScope(name = ApiAccessConstants.JANS_AUTH_CONFIG_READ_ACCESS, description = "View Auth Server properties related information"),
@@ -105,14 +108,15 @@ import jakarta.ws.rs.core.Application;
         @OAuthScope(name = ApiAccessConstants.CLIENT_AUTHORIZATIONS_DELETE_ACCESS, description = "Revoke ClientAuthorizations") }
 
 )))
-public class ApiApplication extends Application {
+public class ApiApplication extends BaseApiApplication {
 
     @Override
     public Set<Class<?>> getClasses() {
         HashSet<Class<?>> classes = new HashSet<>();
 
         // General
-        classes.add(ObjectMapperContextResolver.class);
+        classes = (HashSet) addCommonClasses((classes));
+        
         classes.add(ApiHealthCheck.class);
 
         // oAuth Config