+++ docker-jans-link/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.0.20-1](https://github.com/JanssenProject/jans/compare/docker-jans-link-v1.0.19-1...docker-jans-link-v1.0.20-1) (2023-11-08)
+
+
+### Features
+
+* **docker-all-in-one:** create image requirements ([#6340](https://github.com/JanssenProject/jans/issues/6340)) ([67e8abc](https://github.com/JanssenProject/jans/commit/67e8abc6a3067f4949ea0c4da4d41b370eef53fb))
+* **jans-pycloudlib:** add lock to prevent race condition ([#6329](https://github.com/JanssenProject/jans/issues/6329)) ([9dd82da](https://github.com/JanssenProject/jans/commit/9dd82da5c87ee829c73a1135ce8740b8353f8ab5))
+
+
+### Bug Fixes
+
+* **jans-pycloudlib:** incorrect persistence entry check for ldap and couchbase ([#6297](https://github.com/JanssenProject/jans/issues/6297)) ([87ac453](https://github.com/JanssenProject/jans/commit/87ac453c121b9bdd2dfd7cb4ebb7b8628f322474))
+* prepare for 1.0.20 release ([c6e806e](https://github.com/JanssenProject/jans/commit/c6e806eb31fed998d52cbef7a7d94c231d913102))
+
 ## [1.0.19-1](https://github.com/JanssenProject/jans/compare/docker-jans-link-v1.0.17-1...docker-jans-link-v1.0.19-1) (2023-10-11)